# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Academia Fight Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, teacher, admin]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      required: [id, name, email, role, unit, emailVerified, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        unit:
          type: string
          enum: [matriz, panobianco, upfitness]
        emailVerified:
          type: boolean
        phone:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        paymentDay:
          type: ["integer", "null"]
        modalityThai:
          type: ["boolean", "null"]
        modalityJiu:
          type: ["boolean", "null"]
        thaiGrade:
          type: ["string", "null"]
        thaiGradeColor:
          type: ["string", "null"]
        jiuGrade:
          type: ["string", "null"]
        jiuGradeColor:
          type: ["string", "null"]
        jiuDegree:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
          enum: [matriz, panobianco, upfitness]
          nullable: true
        phone:
          type: string
        profilePhotoUrl:
          type: string
        birthDate:
          type: string
          nullable: true
        paymentDay:
          type: integer
          nullable: true
        modalityThai:
          type: boolean
          nullable: true
        modalityJiu:
          type: boolean
          nullable: true
        thaiGrade:
          type: string
          nullable: true
        thaiGradeColor:
          type: string
          nullable: true
        jiuGrade:
          type: string
          nullable: true
        jiuGradeColor:
          type: string
          nullable: true
        jiuDegree:
          type: integer
          nullable: true
...